Rename .h files to .hpp - Fixes issue #36
