+++ captived/src/http/http_server.inl
@@ -5,8 +5,8 @@
 #include <iostream>
 #include <unistd.h>
 
-#include "defines.h"
-#include "json.h"
+#include "defines.hpp"
+#include "json.hpp"
 
 namespace captiverc {
