+++ captived/src/rest/get_file.cpp
@@ -2,8 +2,8 @@
 #include <fstream>
 #include <sstream>
 
-#include "rest/resource.h"
-#include "rest/get_file.h"
+#include "rest/resource.hpp"
+#include "rest/get_file.hpp"
 
 namespace captiverc {
