+++ captived/src/rest/mode_get_put.cpp
@@ -2,9 +2,9 @@
 #include <fstream>
 #include <sstream>
 
-#include "rest/resource.h"
-#include "rest/get_file.h"
-#include "rest/mode_get_put.h"
+#include "rest/resource.hpp"
+#include "rest/get_file.hpp"
+#include "rest/mode_get_put.hpp"
 
 namespace captiverc {
