+++ captived/src/rest/root_resource.hpp
@@ -5,7 +5,7 @@
 #include <event2/buffer.h>
 #include <event2/http.h>
 
-#include "rest/resource.h"
+#include "rest/resource.hpp"
 
 namespace captiverc {
